rename lock mods
